Changed bugtracker system to a config option.
In the condig file we had sections like:

    jira:
      url: https://something.atlassian.net
      username: admin
      password: secret
      projects: "proj"

"jira" duplicated as a label for the config section and as an
identifier for which bugtracker system we used.

This has been changed so that the bugtracker system is defined in an
option in the section and the section is a pure label. This allows us
to reference more than one bugtracker per bugtrackung system. I.e.

    something:
      bugtracker: jira
      url: https://something.atlassian.net
      username: admin
      password: secret
      projects: "proj"

    another:
      bugtracker: jira
      url: https://something-else.atlassian.net
      username: admin
      password: secret
      projects: "proj"

As a fallback for old configs the label will be used as indicator of
bugtracking system if no bugtracker is configured explicitly.