Refactor API initialization
The getApi() method makes little sense - it does 
not return the API class and the API should not be 
accessed outside of a bugtracker instance.

The bugtracker configuration is passed to the 
object constructor and for JIRA and FogBugz we 
initialize the API here.

Having the configuration available within the
constructor also allows for other refactorings.